xpserver WIN

+++ xpserver/SDK/CHeaders/Widgets/XPWidgetDefs.h
@@ -2,11 +2,11 @@
 #define _XPWidgetDefs_h_
 
 /*
- * Copyright 2005-2012 Sandy Barbour and Ben Supnik
+ * Copyright 2005 Sandy Barbour and Ben Supnik
  * 
  * All rights reserved.  See license.txt for usage.
  * 
- * X-Plane SDK Version: 2.1.1                                                  
+ * X-Plane SDK Version: 1.0.2                                                  
  *
  */
 
@@ -82,7 +82,7 @@ typedef void * XPWidgetID;
  * XPWidgetPropertyID
  * 
  * Properties are values attached to instances of your widgets.  A property is 
- * identified by a 32-bit ID and its value is the width of a pointer.   
+ * identified by a 32-bit ID and its value is also 32 bits.   
  * 
  * Each widget instance may have a property or not have it.  When you set a 
  * property on a widget for the first time, the property is added to the 
@@ -126,7 +126,7 @@ enum {
 
 
 };
-typedef int XPWidgetPropertyID;
+typedef long XPWidgetPropertyID;
 
 /*
  * XPMouseState_t
@@ -208,7 +208,7 @@ enum {
 
 
 };
-typedef int XPDispatchMode;
+typedef long XPDispatchMode;
 
 /*
  * XPWidgetClass
@@ -218,7 +218,7 @@ typedef int XPDispatchMode;
  * Most widgets can be made right from classes.                                
  *
  */
-typedef int XPWidgetClass;
+typedef long XPWidgetClass;
 
 /* An unspecified widget class.  Other widget classes are in                   *
  * XPStandardWidgets.h                                                         */
@@ -451,7 +451,7 @@ enum {
 
 
 };
-typedef int XPWidgetMessage;
+typedef long XPWidgetMessage;
 
 /***************************************************************************
  * WIDGET CALLBACK FUNCTION
@@ -467,7 +467,7 @@ typedef int XPWidgetMessage;
  * 
  * This function defines your custom widget's behavior.  It will be called by 
  * the widgets library to send messages to your widget.  The message and 
- * widget ID are passed in, as well as two ptr-width signed parameters whose 
+ * widget ID are passed in, as well as two 32-bit signed parameters whose 
  * meaning varies with the message.  Return 1 to indicate that you have 
  * processed the message, 0 to indicate that you have not.  For any message 
  * that is not understood, return 0.                                           
@@ -476,8 +476,8 @@ typedef int XPWidgetMessage;
 typedef int (* XPWidgetFunc_t)(
                                    XPWidgetMessage      inMessage,    
                                    XPWidgetID           inWidget,    
-                                   intptr_t             inParam1,    
-                                   intptr_t             inParam2);    
+                                   long                 inParam1,    
+                                   long                 inParam2);    
 
 #ifdef __cplusplus
 }
