+++ xpserver/SDK/CHeaders/XPLM/XPLMNavigation.h
@@ -2,11 +2,11 @@
 #define _XPLMNavigation_h_
 
 /*
- * Copyright 2005-2012 Sandy Barbour and Ben Supnik
+ * Copyright 2005 Sandy Barbour and Ben Supnik
  * 
  * All rights reserved.  See license.txt for usage.
  * 
- * X-Plane SDK Version: 2.1.1                                                  
+ * X-Plane SDK Version: 1.0.2                                                  
  *
  */
 
@@ -206,12 +206,7 @@ XPLM_API XPLMNavRef           XPLMFindNavAid(
  * The buffer for IDs should be at least 6 chars and the buffer for names 
  * should be at least 41 chars, but since these values are likely to go up, I 
  * recommend passing at least 32 chars for IDs and 256 chars for names when 
- * possible. 
- * 
- * The outReg parameter tells if the navaid is within the local "region" of 
- * loaded DSFs.  (This information may not be particularly useful to plugins.) 
- * The parameter is a single byte value 1 for true or 0 for false, not a C 
- * string.                                                                     
+ * possible.                                                                   
  *
  */
 XPLM_API void                 XPLMGetNavAidInfo(
@@ -248,7 +243,7 @@ XPLM_API void                 XPLMGetNavAidInfo(
  * This routine returns the number of entries in the FMS.                      
  *
  */
-XPLM_API int                  XPLMCountFMSEntries(void);
+XPLM_API long                 XPLMCountFMSEntries(void);
 
 /*
  * XPLMGetDisplayedFMSEntry
@@ -256,7 +251,7 @@ XPLM_API int                  XPLMCountFMSEntries(void);
  * This routine returns the index of the entry the pilot is viewing.           
  *
  */
-XPLM_API int                  XPLMGetDisplayedFMSEntry(void);
+XPLM_API long                 XPLMGetDisplayedFMSEntry(void);
 
 /*
  * XPLMGetDestinationFMSEntry
@@ -264,7 +259,7 @@ XPLM_API int                  XPLMGetDisplayedFMSEntry(void);
  * This routine returns the index of the entry the FMS is flying to.           
  *
  */
-XPLM_API int                  XPLMGetDestinationFMSEntry(void);
+XPLM_API long                 XPLMGetDestinationFMSEntry(void);
 
 /*
  * XPLMSetDisplayedFMSEntry
@@ -273,7 +268,7 @@ XPLM_API int                  XPLMGetDestinationFMSEntry(void);
  *  *
  */
 XPLM_API void                 XPLMSetDisplayedFMSEntry(
-                                   int                  inIndex);    
+                                   long                 inIndex);    
 
 /*
  * XPLMSetDestinationFMSEntry
@@ -282,7 +277,7 @@ XPLM_API void                 XPLMSetDisplayedFMSEntry(
  *
  */
 XPLM_API void                 XPLMSetDestinationFMSEntry(
-                                   int                  inIndex);    
+                                   long                 inIndex);    
 
 /*
  * XPLMGetFMSEntryInfo
@@ -297,11 +292,11 @@ XPLM_API void                 XPLMSetDestinationFMSEntry(
  *
  */
 XPLM_API void                 XPLMGetFMSEntryInfo(
-                                   int                  inIndex,    
+                                   long                 inIndex,    
                                    XPLMNavType *        outType,    /* Can be NULL */
                                    char *               outID,    /* Can be NULL */
                                    XPLMNavRef *         outRef,    /* Can be NULL */
-                                   int *                outAltitude,    /* Can be NULL */
+                                   long *               outAltitude,    /* Can be NULL */
                                    float *              outLat,    /* Can be NULL */
                                    float *              outLon);    /* Can be NULL */
 
@@ -315,9 +310,9 @@ XPLM_API void                 XPLMGetFMSEntryInfo(
  *
  */
 XPLM_API void                 XPLMSetFMSEntryInfo(
-                                   int                  inIndex,    
+                                   long                 inIndex,    
                                    XPLMNavRef           inRef,    
-                                   int                  inAltitude);    
+                                   long                 inAltitude);    
 
 /*
  * XPLMSetFMSEntryLatLon
@@ -327,10 +322,10 @@ XPLM_API void                 XPLMSetFMSEntryInfo(
  *
  */
 XPLM_API void                 XPLMSetFMSEntryLatLon(
-                                   int                  inIndex,    
+                                   long                 inIndex,    
                                    float                inLat,    
                                    float                inLon,    
-                                   int                  inAltitude);    
+                                   long                 inAltitude);    
 
 /*
  * XPLMClearFMSEntry
@@ -340,7 +335,7 @@ XPLM_API void                 XPLMSetFMSEntryLatLon(
  *
  */
 XPLM_API void                 XPLMClearFMSEntry(
-                                   int                  inIndex);    
+                                   long                 inIndex);    
 
 /***************************************************************************
  * GPS RECEIVER
